pyproject.toml
[project]
name = "promocode-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.17.1",
    "ruff>=0.15.10",
]
